win-build workflow enhancement.

+++ build.ps1
@@ -6,11 +6,13 @@ pyinstaller --clean -y --onefile main.py
 
 mv .\dist\main.exe .\run.exe
 
-$compress = @{
-  Path = ".\run.exe", ".\webview\", ".\LICENSE", ".\README.md"
-  CompressionLevel = "Fastest"
-  DestinationPath = ".\ms-forms-history-extractor"
-}
-Compress-Archive @compress
+if ($args[0] -ne "CI") {
+	$compress = @{
+	Path = ".\run.exe", ".\webview\", ".\LICENSE", ".\README.md"
+	CompressionLevel = "Fastest"
+	DestinationPath = ".\ms-forms-history-extractor"
+	}
+	Compress-Archive @compress
 
-rm -r .\__pycache__, .\build, .\dist, .\main.spec, .\run.exe
+	rm -r .\__pycache__, .\build, .\dist, .\main.spec, .\run.exe
+}
